Fix empty REPO variable
Since we use multi-pipeline `REPO` variable
was empty because the `CI_REPOSITORY_URL` contains a token
and gitlab does not propagate sensible information.

So, we set `REPO` with `CI_PROJECT_URL` because the project releases
is public and we can use this url to clone the project.

Fix: https://gitlab.e.foundation/e/backlog/-/issues/4553